Described the `operation` concept.
I now await the howls of derision. But this is just a glorified
macro assembler they will scream, and truly they will have
understood my goal: to recreate and reinvent most every part of
computer science conceived in the 60s. That's the 1960s by the way.
How scary it is to think we are now close to the 2060s than the
1960s.